% comment
\name{getArguments()}
\title{Gets arguments from command line.}
\usage{
  args <- getArguments();
}
\arguments{ <none> }
\value{ A vector containing the command line arguments.}
\description{ Get arguments from the command line.}

\source{
   \url{https://www.github.com/FrankThomasTveter/metfark/farkRPackage}
}
\keyword{fark}
