% comment
\name{norwegian(data,idcol,ll)}
\title{Create selection vector for norwegian stations.}
\usage{
  nn <- norwegian(data,idcol,ll);
}
\arguments{
  \describe{
      \item{data}{Matrix.}
      \item{idcol}{Name of station identity column in data.}
      \item{ll}{Optional: Old selection vector.}
   }
}
\value{Returns the selection vector for norwegian stations.}
\description{ Create selection vector for norwegian stations.}

\source{
   \url{https://www.github.com/FrankThomasTveter/metfark/farkRPackage}
}
\keyword{fark}
